♻️ refactor: api 및 파일경로/ 이름 수정
